+++ commands/actions.go
@@ -2,6 +2,7 @@ package commands
 
 import (
 	"fmt"
+	"reflect"
 
 	"github.com/leanovate/gopter"
 	"github.com/leanovate/gopter/gen"
@@ -38,23 +39,23 @@ type sizedCommands struct {
 
 func actionsShrinker(v interface{}) gopter.Shrink {
 	a := v.(*actions)
-	return gen.SliceShrinker(gopter.NoShrinker)(a.sequentialCommands).Map(func(v interface{}) interface{} {
+	return gen.SliceShrinker(gopter.NoShrinker)(a.sequentialCommands).Map(func(v []Command) *actions {
 		return &actions{
 			initialState:       a.initialState,
-			sequentialCommands: v.([]Command),
+			sequentialCommands: v,
 		}
 	})
 }
 
 func genActions(commands Commands) gopter.Gen {
 	return commands.GenInitialState().FlatMap(func(initialState interface{}) gopter.Gen {
-		return genSizedCommands(commands, initialState.(State)).Map(func(v interface{}) interface{} {
+		return genSizedCommands(commands, initialState.(State)).Map(func(v sizedCommands) *actions {
 			return &actions{
 				initialState:       initialState.(State),
-				sequentialCommands: v.(sizedCommands).commands,
+				sequentialCommands: v.commands,
 			}
 		}).WithShrinker(actionsShrinker)
-	})
+	}, reflect.TypeOf((*actions)(nil)))
 }
 
 func genSizedCommands(commands Commands, inistialState State) gopter.Gen {
@@ -66,15 +67,15 @@ func genSizedCommands(commands Commands, inistialState State) gopter.Gen {
 		for i := 0; i < genParams.Size; i++ {
 			gen = gen.FlatMap(func(v interface{}) gopter.Gen {
 				prev := v.(sizedCommands)
-				return commands.GenCommand(prev.state).SuchThat(func(command interface{}) bool {
-					return command.(Command).PreCondition(prev.state)
-				}).Map(func(command interface{}) interface{} {
+				return commands.GenCommand(prev.state).SuchThat(func(command Command) bool {
+					return command.PreCondition(prev.state)
+				}).Map(func(command Command) sizedCommands {
 					return sizedCommands{
-						state:    command.(Command).NextState(prev.state),
-						commands: append(prev.commands, command.(Command)),
+						state:    command.NextState(prev.state),
+						commands: append(prev.commands, command),
 					}
 				})
-			})
+			}, reflect.TypeOf(sizedCommands{}))
 		}
 		return gen(genParams)
 	}