+++ gen/ptr_of_test.go
@@ -48,11 +48,7 @@ func TestPtrOfFoo(t *testing.T) {
 }
 
 func GenFoo() gopter.Gen {
-	return gen.SliceOfN(16, gen.Rune()).Map(func(v interface{}) interface{} {
-		asRunes, ok := v.([]rune)
-		if !ok {
-			panic("Oh craps")
-		}
-		return Foo(asRunes)
+	return gen.SliceOfN(16, gen.Rune()).Map(func(v []rune) Foo {
+		return Foo(v)
 	})
 }