+++ gen_test.go
@@ -24,8 +24,8 @@ func TestGenSample(t *testing.T) {
 
 func TestGenMap(t *testing.T) {
 	gen := constGen("sample")
-	var mappedWith interface{}
-	mapper := func(v interface{}) interface{} {
+	var mappedWith string
+	mapper := func(v string) string {
 		mappedWith = v
 		return "other"
 	}
@@ -33,7 +33,7 @@ func TestGenMap(t *testing.T) {
 	if !ok || value != "other" {
 		t.Errorf("Invalid gen sample: %#v", value)
 	}
-	if mappedWith.(string) != "sample" {
+	if mappedWith != "sample" {
 		t.Errorf("Invalid mapped with: %#v", mappedWith)
 	}
 
@@ -46,6 +46,32 @@ func TestGenMap(t *testing.T) {
 	}
 }
 
+func TestGenMapNoFunc(t *testing.T) {
+	defer expectPanic(t, "Param of Map has to be a func, but is string")
+	constGen("sample").Map("not a function")
+}
+
+func TestGenMapTooManyParams(t *testing.T) {
+	defer expectPanic(t, "Param of Map has to be a func with one param, but is 2")
+	constGen("sample").Map(func(a, b string) string {
+		return ""
+	})
+}
+
+func TestGenMapToInvalidParamtype(t *testing.T) {
+	defer expectPanic(t, "Param of Map has to be a func with one param assignable to string, but is int")
+	constGen("sample").Map(func(a int) string {
+		return ""
+	})
+}
+
+func TestGenMapToManyReturns(t *testing.T) {
+	defer expectPanic(t, "Param of Map has to be a func with one return value, but is 2")
+	constGen("sample").Map(func(a string) (string, bool) {
+		return "", false
+	})
+}
+
 func TestGenFlatMap(t *testing.T) {
 	gen := constGen("sample")
 	var mappedWith interface{}
@@ -53,7 +79,7 @@ func TestGenFlatMap(t *testing.T) {
 		mappedWith = v
 		return constGen("other")
 	}
-	value, ok := gen.FlatMap(mapper).Sample()
+	value, ok := gen.FlatMap(mapper, reflect.TypeOf("")).Sample()
 	if !ok || value != "other" {
 		t.Errorf("Invalid gen sample: %#v", value)
 	}
@@ -64,7 +90,7 @@ func TestGenFlatMap(t *testing.T) {
 	gen = gen.SuchThat(func(interface{}) bool {
 		return false
 	})
-	value, ok = gen.FlatMap(mapper).Sample()
+	value, ok = gen.FlatMap(mapper, reflect.TypeOf("")).Sample()
 	if ok {
 		t.Errorf("Invalid gen sample: %#v", value)
 	}
@@ -96,8 +122,8 @@ func TestCombineGens(t *testing.T) {
 }
 
 func TestSuchThat(t *testing.T) {
-	var sieveArg interface{}
-	sieve := func(v interface{}) bool {
+	var sieveArg string
+	sieve := func(v string) bool {
 		sieveArg = v
 		return true
 	}
@@ -110,9 +136,9 @@ func TestSuchThat(t *testing.T) {
 		t.Errorf("Invalid sieveArg: %#v", sieveArg)
 	}
 
-	sieveArg = nil
-	var sieve2Arg interface{}
-	sieve2 := func(v interface{}) bool {
+	sieveArg = ""
+	var sieve2Arg string
+	sieve2 := func(v string) bool {
 		sieve2Arg = v
 		return false
 	}
@@ -127,6 +153,39 @@ func TestSuchThat(t *testing.T) {
 	if sieve2Arg != "sample" {
 		t.Errorf("Invalid sieve2Arg: %#v", sieve2Arg)
 	}
+}
+
+func TestGenSuchThatNoFunc(t *testing.T) {
+	defer expectPanic(t, "Param of SuchThat has to be a func, but is string")
+	constGen("sample").SuchThat("not a function")
+}
+
+func TestGenSuchTooManyParams(t *testing.T) {
+	defer expectPanic(t, "Param of SuchThat has to be a func with one param, but is 2")
+	constGen("sample").SuchThat(func(a, b string) bool {
+		return false
+	})
+}
+
+func TestGenSuchThatToInvalidParamtype(t *testing.T) {
+	defer expectPanic(t, "Param of SuchThat has to be a func with one param assignable to string, but is int")
+	constGen("sample").SuchThat(func(a int) bool {
+		return false
+	})
+}
+
+func TestGenSuchToManyReturns(t *testing.T) {
+	defer expectPanic(t, "Param of SuchThat has to be a func with one return value, but is 2")
+	constGen("sample").SuchThat(func(a string) (string, bool) {
+		return "", false
+	})
+}
+
+func TestGenSuchToInvalidReturns(t *testing.T) {
+	defer expectPanic(t, "Param of SuchThat has to be a func with one return value of bool, but is string")
+	constGen("sample").SuchThat(func(a string) string {
+		return ""
+	})
 }
 
 func TestWithShrinker(t *testing.T) {
@@ -144,5 +203,14 @@ func TestWithShrinker(t *testing.T) {
 	result.Shrinker(value)
 	if shrinkerArg != "sample" {
 		t.Errorf("Invalid shrinkerArg: %#v", shrinkerArg)
+	}
+}
+
+func expectPanic(t *testing.T, expected string) {
+	r := recover()
+	if r == nil {
+		t.Errorf("The code did not panic")
+	} else if r.(string) != expected {
+		t.Errorf("Panic does not match: '%#v' != '%#v'", r, expected)
 	}
 }